backend/requirements.txt
flask
flask-cors
flask_restful
python-dateutil
requests
reportlab
matplotlib
faker 